Programming Assignments/PA2_2/requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
tqdm>=4.66.0
PyYAML>=6.0
pandas>=2.0.0
matplotlib>=3.8.0
numpy>=1.24.0
scipy>=1.11.0
sentencepiece>=0.2.0
protobuf>=3.20.0
huggingface_hub>=0.22.0
